backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
tensorflow-cpu==2.15.0
pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
